Contracts v9 (#2256)
* Protection against front-running and reentrancy

Adding reentrancy guard on `activateExodusMode()`.
Checking the length of block and proof.commitments in `proveBlocks(...)`

* Protection against reentrancy attack

Adding reentrancy guard on all external functions that write to storage.

* Update ZkSync.sol

Protection against duplicated proved blocks.

* Update ZkSync.sol

Adding a comment for front-running protection.

* Update ZkSync.sol

Gas Optimization

* protection against griefing attack

So, the server can check it's value before adding the `changePubKey` to the block (to protect against griefing attack).

* Update AdditionalZkSyncCutNoticePeriodUnitTest.sol

Initializing reentrancy guard.

* immunefi bugs

* removing complete withdrawals

* removing complete withdrawals

* removing complete withdrawals

* Add server protection of block invalidation

* Try to fix unit tests

* Restore tests

* Try to find workaround

* Fix unit tests

* gas optimization

* gas optimization & change governor check

* increasing the code readability

* solving the mistake

* fixing the logic break due to gas optimization

* Update contracts/contracts/ZkSync.sol


* cross-domain attck protection

* domain-typehash changed

* new change pubkey for cross-domain protection

* removing extra spaces

* adding test for changePubKey type EIP712

* name struct with a capital letter

* prove blocks unit test + memory to calldata + adding length check to proveBlocks function

* removing the redundant function

* Fix ci


* Update ZkSync.sol

Changing the require message to be unique from "r" to "o3".

* updating the front-running protection

removing the proof.commitments[] length check.

* Adding comment

Adding comment to the require in `proveBlocks` function

* Fix comment

* Add primitives


* Implement change pubkey eip712


* Use eip712 change pub key in zksync-rs


* Add test


* Deploy


* Fix


* Add eip712 to js sdk


* Some refactoring


* Add check eth witness generation


* Fix sdk


* Fix eth witness generation


* Fix types


* Remove strict node js


* Fix ecdsa signatures


* Add assertion for const generic size


* Use u32 for chain id


* Fix comment about costs for change pub key


* Fix some nits


* Change name of deprecated change pub key method


* Smart contracts v9 (last changes before upgrade) (#2285)

* Add event for different type of withdrawals

* Fix bug from compiler side

* Add owner field into withdraw event

* Fix networks



* Fix typos

* Add a changelog

* Small fix

* Add missing point

* Add info about changed event signature

* Merge  (#2295)

* Remove sepolia (#2291)



* Add correct pagination for transaction history query


* Fix transactions_history_item_old


* Fix tests but a lot of cludges


* Fix test

+++ core/lib/config/src/configs/api.rs
@@ -230,7 +230,7 @@ pub struct Web3Config {
     pub url: String,
     /// Max difference between blocks in `eth_getLogs` method.
     pub max_block_range: u32,
-    pub chain_id: u32,
+    pub chain_id: u64,
 }
 
 impl Web3Config {
